requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
flask==2.3.3
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-mail==0.9.1
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
multidict==6.0.4
pyjwt==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.22
twilio==8.10.0
typing_extensions==4.8.0
urllib3==2.0.7
werkzeug==2.3.7
wtforms==3.1.0
yarl==1.9.2
